% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{clr_min_percent}
\alias{clr_min_percent}
\title{CLR (centered log-ratio) transformation of phyloseq object (pseudocount equal to minimum_value/2).}
\usage{
clr_min_percent(physeq)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}
}
\value{
A phyloseq object with transformed abundance data.
}
\description{
This function performs the Chord transformation on the abundance data of a phyloseq object.
The pseudocount used to avoid 0s when log transforming is equal to the minimum abundance value divided by 2.
This transformation should be used with raw abundance data (internal Total Sum Scaling transformation -- relative abundances).
}
\examples{
data("GlobalPatterns")
transformed.data <- clr_min_percent(GlobalPatterns)

}
