% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Files2Phyloseq.R
\name{Files2Phyloseq}
\alias{Files2Phyloseq}
\title{Files2Phyloseq}
\usage{
Files2Phyloseq(
  BIOM.file = NULL,
  OTU.file = NULL,
  TAXA.file = NULL,
  META.file = NULL,
  TREE.file = NULL,
  unclass.file = NULL,
  mphlan.file = NULL,
  humann.file = NULL,
  sep = "\\t",
  str2rm = "",
  factor.lev = NULL,
  filt.var = NULL,
  lev.kept = NULL
)
}
\arguments{
\item{BIOM.file}{(Required if not providing OTU and TAXA files) A character string of the PATH to the BIOM file}

\item{OTU.file}{(Required if not providing BIOM file) A character string of the PATH to the ASV/OTU/Species abundance table file
(accepted formats: qza, csv, tsv and txt if "sep" is specified).}

\item{TAXA.file}{(Required if not providing BIOM file) A character string of the PATH to the taxonomic classification file
(accepted formats: qza, csv, tsv and txt if "sep" is specified).}

\item{META.file}{(Optional) A character string of the PATH to the metadata file
(accepted formats: xls, xlsx, xlsm, csv, tsv and txt if "sep" is specified).
The first column must be named "SampleID" and must contain the sample names.}

\item{TREE.file}{(Optional) A character string of the PATH to the phylogenetic tree file
(accepted formats: qza and tree).}

\item{unclass.file}{(Optional) A character string of the PATH to the unclassified file.
When using Bracken the "Unclassified" taxon is removed, to add this information extract it from the Kraken report and create a txt file
with the the sample names in the first column and the corresponding "Unclassified" values in the second column.
Use the Extract_unclassified_reads.sh function to do this.
(accepted formats: tsv)}

\item{mphlan.file}{(Optional) A character string of the PATH to the MetaPhlAn3 file.
(accepted formats: tsv)}

\item{humann.file}{(Optional) A character string of the PATH to the HUMAnN3 file.
(accepted formats: tsv)}

\item{sep}{(Optional)     A character string of the field separator. Values on each line of the file are separated by this character.
Only required if the OTU, TAXA or META files are in "txt" format. DEFAULT: "\t"}

\item{str2rm}{(Optional) A vector of character strings defining the patterns to be removed from the names of the samples. DEFAULT: ""}

\item{factor.lev}{(Optional) A list containing the names of the levels found in each categorical variable.
This will establish in which order level names will be displayed in the plots. DEFAULT: NULL}

\item{filt.var}{(Optional) A character string of the metadata variable to use for the filtration of samples. DEFAULT: NULL}

\item{lev.kept}{(Optional) A vector containing the names of the levels from the 'filt.var' that will be retained. DEFAULT: NULL}
}
\value{
A phyloseq object in which the variables defined as factors have their levels organised in the specified order.
}
\description{
This function is used to convert DADA2, QIIME2, Kraken2, Bracken, MetaPhlAn3 and HUMAnN3 output files into a phyloseq object.
It accepts multiple file formats: qiime artifact - "qza", comma-separated - "csv", tab-separated - "tsv",
Newick-format - "tree" , Excel - "xls, xlsx, xlsm" and text - "txt. If file format is "txt" please specify field separator.
Filtration of specific samples based on metadata variables is also possible.
}
\examples{
To create a phyloseq object from the DADA2 output files:
Files2Phyloseq("~/DADA2/seqtab.csv", "~/DADA2/taxaId.csv", "~/DADA2/Metadata.xlsm", factor.lev = list(Type = c("Water", "Sediment"), Source = c("Market", "Wild" )))

To create a phyloseq object from the Bracken output files:
Files2Phyloseq("~/DADA2/seqtab.csv", "~/DADA2/taxaId.csv", "~/DADA2/Metadata.xlsm", factor.lev = list(Type = c("Water", "Sediment"), Source = c("Market", "Wild" )))
}
