% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{logchord_0_plus_1}
\alias{logchord_0_plus_1}
\title{Chord transformation of phyloseq object (pseudocount of 1 added only to values = 0).}
\usage{
logchord_0_plus_1(physeq)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}
}
\value{
A phyloseq object with transformed abundance data.
}
\description{
This function performs the Chord transformation on the abundance data of a phyloseq object.
The pseudocount used to avoid 0s when log transforming is 1, however, it is only added to values equal to 0.
This transformation should be used with raw abundance data.
}
\examples{
data("GlobalPatterns")
transformed.data <- logchord_0_plus_1(GlobalPatterns)

}
