% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{wrap.strip.colors}
\alias{wrap.strip.colors}
\title{# Add colors to ggplot wrap strips}
\usage{
wrap.strip.colors(
  df,
  plot,
  Top.var = NULL,
  Bot.var = NULL,
  Top.strip.fill = "by.Top.var",
  Bot.strip.fill = "by.Bot.var",
  col.palette = c("#ff0000", "#00ff00", "#0000ff"),
  display = TRUE
)
}
\arguments{
\item{df}{(Required) Data frame used to create the existing plot.}

\item{plot}{(Required) Existing plot created using the facet_wrap option.}

\item{Top.var}{(Required) A character string indicating the data frame variable used in facet_wrap. DEFAULT: NULL}

\item{Bot.var}{(Optional) A character string indicating a second data frame variable used in facet_wrap. DEFAULT: NULL}

\item{Top.strip.fill}{(Optional) A vector with the hexadecimal colors associated to the levels of the Top variable.
OR a character string ("by.Top.var" or "by.Bot.var" or "blank") indicating if colors are assigned to the levels of the Top or Bottom variables. DEFAULT: "by.Top.var"}

\item{Bot.strip.fill}{(Optional) A vector with the hexadecimal colors associated to the levels of the Bottom variable.
OR a character string ("by.Top.var", "by.Bot.var" or "blank") indicating if colors are assigned to the levels of the Top or Bottom variables. DEFAULT: "by.Bot.var"}

\item{col.palette}{(Optional) A character vector containing the hexadecimal representations of one or more colors (seedcolors). DEFAULT: c("#ff0000", "#00ff00", "#0000ff").}

\item{display}{(Optional) A logical value. If TRUE the function returns a ggplot object, otherwise, it returns a gtable() object. DEFAULT: TRUE}
}
\value{
A ggplot object or if display = FALSE a gtable() object.
}
\description{
This function adds random or specified colors to the strips of panels created by facet_wrap.
Colors can be added to the Top and/or Bottom strips. They also can be assigned to the levels of the variables selected.
}
\examples{
data("GlobalPatterns")
colnames(tax_table(GlobalPatterns)) <- c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
filt.data <- FilterAndSummarise(GlobalPatterns)
mdf = psmelt(filt.data)
p = ggplot(mdf, aes(x=SampleType, y=Abundance, fill=Genus))
p = p + geom_bar(color="black", stat="identity", position="stack") + facet_wrap(~SampleType) + theme(legend.position = "none")
print(p)
p2 <- cowplot::plot_grid(wrap.strip.colors(mdf, p, Top.var = "SampleType", Top.strip.fill = "by.Top.var"))
print(p2)

}
